Increase Java language version to Java 8
It's 2025, and Java 7 is thoroughly dead at this point. Java 8 is still
hanging on for legacy support, which I guess this is...

In the past, opsu! has crashed when I tried to run it on Java versions
newer than 8, so I'm stuck with that for now. However, I am interested
in the new Foreign Function and Memory API, so I probably won't stay on
Java 8 forever.