Actually load Zig code
Previously, the Zig code wasn't being loaded at all by the Java code.

Right now, the loading is done through the Miniaudio class. This is fine
while miniaudio is the only thing the native code is used for, but if it
starts being used for more things we'll probably want to change that.